agents/options-trading/requirements.txt
yfinance>=0.2.0
numpy>=1.20.0
pandas>=1.3.0
python-dotenv>=0.19.0
aiohttp>=3.8.0